feat(code-editor): add support for line wrapping when passing prop
